Avoid star imports for PortingMessage
