+++ MODULEINFO
@@ -54,6 +54,7 @@ Make: dmlc
     $(HOST)/bin/dml/python/dml/messages.py
     $(HOST)/bin/dml/python/dml/objects.py
     $(HOST)/bin/dml/python/dml/output.py
+    $(HOST)/bin/dml/python/dml/porting.py
     $(HOST)/bin/dml/python/dml/provisional.py
     $(HOST)/bin/dml/python/dml/reginfo.py
     $(HOST)/bin/dml/python/dml/serialize.py
